Rename `Analyze` -> `Eval`.
