+++ pact.cabal
@@ -55,8 +55,13 @@ library
   if !impl(ghcjs)
     hs-source-dirs: src-ghc
     exposed-modules:
-      Pact.Analyze.Analyze
-      Pact.Analyze.AnalyzeNumerical
+      Pact.Analyze.Eval
+      Pact.Analyze.Eval.Invariant
+      Pact.Analyze.Eval.Numerical
+      Pact.Analyze.Eval.Prop
+      Pact.Analyze.Eval.Pure
+      Pact.Analyze.Eval.Term
+      Pact.Analyze.Eval.Types
       Pact.Analyze.Check
       Pact.Analyze.Errors
       Pact.Analyze.LegacySFunArray
