+++ src/Pact/Analyze/Eval/Types.hs
@@ -1,0 +1,463 @@
+{-# LANGUAGE FlexibleContexts           #-}
+{-# LANGUAGE GeneralizedNewtypeDeriving #-}
+{-# LANGUAGE MultiParamTypeClasses      #-}
+{-# LANGUAGE NamedFieldPuns             #-}
+{-# LANGUAGE OverloadedStrings          #-}
+{-# LANGUAGE Rank2Types                 #-}
+{-# LANGUAGE ScopedTypeVariables        #-}
+{-# LANGUAGE TemplateHaskell            #-}
+{-# LANGUAGE TypeFamilies               #-}
+{-# LANGUAGE TypeOperators              #-}
+module Pact.Analyze.Eval.Types where
+
+import           Control.Applicative          (ZipList (..))
+import           Control.Lens                 (Lens', at, ifoldl, ix, lens,
+                                               makeLenses, singular, view, (&),
+                                               (<&>), (?~), _2)
+import           Control.Monad.Except         (MonadError)
+import           Control.Monad.Reader         (MonadReader)
+import           Data.Map.Strict              (Map)
+import qualified Data.Map.Strict              as Map
+import           Data.Maybe                   (mapMaybe)
+import           Data.Monoid                  ((<>))
+import           Data.SBV                     (Boolean (true), HasKind,
+                                               Mergeable (symbolicMerge), SBV,
+                                               SBool,
+                                               SymArray (readArray, writeArray),
+                                               SymWord, Symbolic, false,
+                                               uninterpret)
+import qualified Data.SBV.Internals           as SBVI
+import           Data.Text                    (Text)
+import qualified Data.Text                    as T
+
+import           Pact.Types.Lang              (Info)
+import           Pact.Types.Runtime           (PrimType (TyBool, TyDecimal, TyInteger, TyKeySet, TyString, TyTime),
+                                               Type (TyPrim))
+import qualified Pact.Types.Runtime           as Pact
+import qualified Pact.Types.Typecheck         as Pact
+
+import           Pact.Analyze.Errors
+import           Pact.Analyze.LegacySFunArray (SFunArray, mkSFunArray)
+import           Pact.Analyze.Orphans         ()
+import           Pact.Analyze.Types           hiding (tableName)
+import qualified Pact.Analyze.Types           as Types
+import           Pact.Analyze.Util
+
+
+class (MonadError AnalyzeFailure m, S :<: TermOf m) => Analyzer m where
+  type TermOf m   :: * -> *
+  eval            :: (Show a, SymWord a) => TermOf m a -> m (S a)
+  evalO           :: TermOf m Object -> m Object
+  throwErrorNoLoc :: AnalyzeFailureNoLoc -> m a
+  getVar          :: VarId -> m (Maybe AVal)
+
+data AnalyzeEnv
+  = AnalyzeEnv
+    { _aeScope     :: !(Map VarId AVal)              -- used as a stack
+    , _aeKeySets   :: !(SFunArray KeySetName KeySet) -- read-only
+    , _aeKsAuths   :: !(SFunArray KeySet Bool)       -- read-only
+    , _invariants  :: !(TableMap [Located (Invariant Bool)])
+    , _aeModelTags :: !ModelTags
+    , _aeInfo      :: !Info
+    }
+  deriving Show
+
+mkAnalyzeEnv :: [Table] -> ModelTags -> Info -> AnalyzeEnv
+mkAnalyzeEnv tables tags info =
+  let keySets'    = mkFreeArray "keySets"
+      keySetAuths = mkFreeArray "keySetAuths"
+
+      invariants' = TableMap $ Map.fromList $ tables <&>
+        \(Table tname _ut someInvariants) ->
+          (TableName (T.unpack tname), someInvariants)
+
+      argMap :: Map VarId AVal
+      argMap = view (located._2._2) <$> _mtArgs tags
+
+  in AnalyzeEnv argMap keySets' keySetAuths invariants' tags info
+
+mkFreeArray :: (SymWord a, HasKind b) => Text -> SFunArray a b
+mkFreeArray = mkSFunArray . uninterpret . T.unpack . sbvIdentifier
+
+sbvIdentifier :: Text -> Text
+sbvIdentifier = T.replace "-" "_"
+
+data QueryEnv
+  = QueryEnv
+    { _qeAnalyzeEnv    :: AnalyzeEnv
+    , _qeAnalyzeState  :: AnalyzeState
+    , _qeAnalyzeResult :: AVal
+    , _qeTableScope    :: Map VarId TableName
+    --
+    -- TODO: implement column quantification. update 'getLitColName', and
+    -- 'evalProp' for @Forall@ and @Exists@ at the same time.
+    --
+    , _qeColumnScope   :: Map VarId ()
+    }
+
+mkQueryEnv :: AnalyzeEnv -> AnalyzeState -> AVal -> QueryEnv
+mkQueryEnv env state result = QueryEnv env state result Map.empty Map.empty
+
+newtype Constraints
+  = Constraints { runConstraints :: Symbolic () }
+
+instance Show Constraints where
+  show _ = "<symbolic>"
+
+instance Monoid Constraints where
+  mempty = Constraints (pure ())
+  mappend (Constraints act1) (Constraints act2) = Constraints $ act1 *> act2
+
+data SymbolicCells
+  = SymbolicCells
+    { _scIntValues     :: ColumnMap (SFunArray RowKey Integer)
+    , _scBoolValues    :: ColumnMap (SFunArray RowKey Bool)
+    , _scStringValues  :: ColumnMap (SFunArray RowKey String)
+    , _scDecimalValues :: ColumnMap (SFunArray RowKey Decimal)
+    , _scTimeValues    :: ColumnMap (SFunArray RowKey Time)
+    , _scKsValues      :: ColumnMap (SFunArray RowKey KeySet)
+    -- TODO: opaque blobs
+    }
+  deriving (Show)
+
+-- Implemented by-hand until 8.2, when we have DerivingStrategies
+instance Mergeable SymbolicCells where
+  symbolicMerge force test
+    (SymbolicCells a b c d e f)
+    (SymbolicCells a' b' c' d' e' f')
+    = SymbolicCells (m a a') (m b b') (m c c') (m d d') (m e e') (m f f')
+    where
+      m :: Mergeable a => ColumnMap a -> ColumnMap a -> ColumnMap a
+      m = symbolicMerge force test
+
+-- Checking state that is split before, and merged after, conditionals.
+data LatticeAnalyzeState
+  = LatticeAnalyzeState
+    { _lasSucceeds            :: SBV Bool
+    --
+    -- TODO: instead of having a single boolean here, we should probably use
+    --       finer-grained tracking, so that we can test whether a single
+    --       invariant is being maintained
+    --
+    , _lasMaintainsInvariants :: TableMap (ZipList (Located (SBV Bool)))
+    , _lasTablesRead          :: SFunArray TableName Bool
+    , _lasTablesWritten       :: SFunArray TableName Bool
+    , _lasIntCellDeltas       :: TableMap (ColumnMap (SFunArray RowKey Integer))
+    , _lasDecCellDeltas       :: TableMap (ColumnMap (SFunArray RowKey Decimal))
+    , _lasIntColumnDeltas     :: TableMap (ColumnMap (S Integer))
+    , _lasDecColumnDeltas     :: TableMap (ColumnMap (S Decimal))
+    , _lasTableCells          :: TableMap SymbolicCells
+    , _lasRowsRead            :: TableMap (SFunArray RowKey Integer)
+    , _lasRowsWritten         :: TableMap (SFunArray RowKey Integer)
+    , _lasCellsEnforced       :: TableMap (ColumnMap (SFunArray RowKey Bool))
+    -- We currently maintain cellsWritten only for deciding whether a cell has
+    -- been "invalidated" for the purposes of keyset enforcement. If a keyset
+    -- has been overwritten and *then* enforced, that does not constitute valid
+    -- enforcement of the keyset.
+    , _lasCellsWritten        :: TableMap (ColumnMap (SFunArray RowKey Bool))
+    }
+  deriving (Show)
+
+-- Implemented by-hand until 8.2, when we have DerivingStrategies
+instance Mergeable LatticeAnalyzeState where
+  symbolicMerge force test
+    (LatticeAnalyzeState
+      success  tsInvariants  tsRead  tsWritten  intCellDeltas  decCellDeltas
+      intColDeltas  decColDeltas  cells  rsRead  rsWritten  csEnforced  csWritten)
+    (LatticeAnalyzeState
+      success' tsInvariants' tsRead' tsWritten' intCellDeltas' decCellDeltas'
+      intColDeltas' decColDeltas' cells' rsRead' rsWritten' csEnforced' csWritten')
+        = LatticeAnalyzeState
+          (symbolicMerge force test success       success')
+          (symbolicMerge force test tsInvariants  tsInvariants')
+          (symbolicMerge force test tsRead        tsRead')
+          (symbolicMerge force test tsWritten     tsWritten')
+          (symbolicMerge force test intCellDeltas intCellDeltas')
+          (symbolicMerge force test decCellDeltas decCellDeltas')
+          (symbolicMerge force test intColDeltas  intColDeltas')
+          (symbolicMerge force test decColDeltas  decColDeltas')
+          (symbolicMerge force test cells         cells')
+          (symbolicMerge force test rsRead        rsRead')
+          (symbolicMerge force test rsWritten     rsWritten')
+          (symbolicMerge force test csEnforced    csEnforced')
+          (symbolicMerge force test csWritten     csWritten')
+
+-- Checking state that is transferred through every computation, in-order.
+data GlobalAnalyzeState
+  = GlobalAnalyzeState
+    { _gasConstraints   :: Constraints          -- we log these a la writer
+    , _gasKsProvenances :: Map TagId Provenance -- added as we accum ks info
+    }
+  deriving (Show)
+
+data AnalyzeState
+  = AnalyzeState
+    { _latticeState :: LatticeAnalyzeState
+    , _globalState  :: GlobalAnalyzeState
+    }
+  deriving (Show)
+
+data AnalysisResult
+  = AnalysisResult
+    { _arProposition   :: SBV Bool
+    , _arKsProvenances :: Map TagId Provenance
+    }
+  deriving (Show)
+
+makeLenses ''AnalyzeEnv
+makeLenses ''AnalyzeState
+makeLenses ''GlobalAnalyzeState
+makeLenses ''LatticeAnalyzeState
+makeLenses ''SymbolicCells
+makeLenses ''AnalysisResult
+makeLenses ''QueryEnv
+
+
+mkInitialAnalyzeState :: [Table] -> AnalyzeState
+mkInitialAnalyzeState tables = AnalyzeState
+    { _latticeState = LatticeAnalyzeState
+        { _lasSucceeds            = true
+        , _lasMaintainsInvariants = mkMaintainsInvariants
+        , _lasTablesRead          = mkSFunArray $ const false
+        , _lasTablesWritten       = mkSFunArray $ const false
+        , _lasIntCellDeltas       = intCellDeltas
+        , _lasDecCellDeltas       = decCellDeltas
+        , _lasIntColumnDeltas     = intColumnDeltas
+        , _lasDecColumnDeltas     = decColumnDeltas
+        , _lasTableCells          = mkSymbolicCells tables
+        , _lasRowsRead            = mkPerTableSFunArray 0
+        , _lasRowsWritten         = mkPerTableSFunArray 0
+        , _lasCellsEnforced       = cellsEnforced
+        , _lasCellsWritten        = cellsWritten
+        }
+    , _globalState = GlobalAnalyzeState
+        { _gasConstraints   = mempty
+        , _gasKsProvenances = mempty
+        }
+    }
+
+  where
+    tableNames :: [TableName]
+    tableNames = map (TableName . T.unpack . view Types.tableName) tables
+
+    intCellDeltas = mkTableColumnMap (== TyPrim TyInteger) (mkSFunArray (const 0))
+    decCellDeltas = mkTableColumnMap (== TyPrim TyDecimal) (mkSFunArray (const 0))
+    intColumnDeltas = mkTableColumnMap (== TyPrim TyInteger) 0
+    decColumnDeltas = mkTableColumnMap (== TyPrim TyDecimal) 0
+    cellsEnforced
+      = mkTableColumnMap (== TyPrim TyKeySet) (mkSFunArray (const false))
+    cellsWritten = mkTableColumnMap (const True) (mkSFunArray (const false))
+
+    mkTableColumnMap
+      :: (Pact.Type Pact.UserType -> Bool) -> a -> TableMap (ColumnMap a)
+    mkTableColumnMap f defValue = TableMap $ Map.fromList $
+      tables <&> \Table { _tableName, _tableType } ->
+        let fields = Pact._utFields _tableType
+            colMap = ColumnMap $ Map.fromList $ flip mapMaybe fields $
+              \(Pact.Arg argName ty _) ->
+                if f ty
+                then Just (ColumnName (T.unpack argName), defValue)
+                else Nothing
+        in (TableName (T.unpack _tableName), colMap)
+
+    mkPerTableSFunArray :: SBV v -> TableMap (SFunArray k v)
+    mkPerTableSFunArray defaultV = TableMap $ Map.fromList $ zip
+      tableNames
+      (repeat $ mkSFunArray $ const defaultV)
+
+    mkMaintainsInvariants = TableMap $ Map.fromList $
+      tables <&> \Table { _tableName, _tableInvariants } ->
+        (TableName (T.unpack _tableName), ZipList $ const true <$$> _tableInvariants)
+
+mkSymbolicCells :: [Table] -> TableMap SymbolicCells
+mkSymbolicCells tables = TableMap $ Map.fromList cellsList
+  where
+    cellsList = tables <&> \Table { _tableName, _tableType = Pact.Schema _ _ fields _ } ->
+      let fields'  = Map.fromList $
+            map (\(Pact.Arg argName ty _i) -> (argName, ty)) fields
+
+      in (TableName (T.unpack _tableName), mkCells _tableName fields')
+
+    mkCells :: Text -> Map Text (Pact.Type Pact.UserType) -> SymbolicCells
+    mkCells tableName fields = ifoldl
+      (\colName cells ty ->
+        let col      = ColumnName $ T.unpack colName
+
+            mkArray :: forall a. HasKind a => SFunArray RowKey a
+            mkArray  = mkFreeArray $ "cells__" <> tableName <> "__" <> colName
+
+        in cells & case ty of
+             TyPrim TyInteger -> scIntValues.at col     ?~ mkArray
+             TyPrim TyBool    -> scBoolValues.at col    ?~ mkArray
+             TyPrim TyDecimal -> scDecimalValues.at col ?~ mkArray
+             TyPrim TyTime    -> scTimeValues.at col    ?~ mkArray
+             TyPrim TyString  -> scStringValues.at col  ?~ mkArray
+             TyPrim TyKeySet  -> scKsValues.at col      ?~ mkArray
+             --
+             -- TODO: we should Left here. this means that mkSymbolicCells and
+             --       mkInitialAnalyzeState should both return Either.
+             --
+             _                -> id
+      )
+      (SymbolicCells mempty mempty mempty mempty mempty mempty)
+      fields
+
+mkSVal :: SBV a -> SBVI.SVal
+mkSVal (SBVI.SBV v) = v
+
+class HasAnalyzeEnv a where
+  {-# MINIMAL analyzeEnv #-}
+  analyzeEnv :: Lens' a AnalyzeEnv
+
+  scope :: Lens' a (Map VarId AVal)
+  scope = analyzeEnv.aeScope
+
+  keySets :: Lens' a (SFunArray KeySetName KeySet)
+  keySets = analyzeEnv.aeKeySets
+
+  ksAuths :: Lens' a (SFunArray KeySet Bool)
+  ksAuths = analyzeEnv.aeKsAuths
+
+instance HasAnalyzeEnv AnalyzeEnv where analyzeEnv = id
+instance HasAnalyzeEnv QueryEnv   where analyzeEnv = qeAnalyzeEnv
+
+succeeds :: Lens' AnalyzeState (S Bool)
+succeeds = latticeState.lasSucceeds.sbv2S
+
+maintainsInvariants :: Lens' AnalyzeState (TableMap (ZipList (Located SBool)))
+maintainsInvariants = latticeState.lasMaintainsInvariants
+
+tableRead :: TableName -> Lens' AnalyzeState (S Bool)
+tableRead tn = latticeState.lasTablesRead.symArrayAt (literalS tn).sbv2S
+
+tableWritten :: TableName -> Lens' AnalyzeState (S Bool)
+tableWritten tn = latticeState.lasTablesWritten.symArrayAt (literalS tn).sbv2S
+
+intCellDelta
+  :: TableName
+  -> ColumnName
+  -> S RowKey
+  -> Lens' AnalyzeState (S Integer)
+intCellDelta tn cn sRk = latticeState.lasIntCellDeltas.singular (ix tn).
+  singular (ix cn).symArrayAt sRk.sbv2S
+
+decCellDelta
+  :: TableName
+  -> ColumnName
+  -> S RowKey
+  -> Lens' AnalyzeState (S Decimal)
+decCellDelta tn cn sRk = latticeState.lasDecCellDeltas.singular (ix tn).
+  singular (ix cn).symArrayAt sRk.sbv2S
+
+intColumnDelta :: TableName -> ColumnName -> Lens' AnalyzeState (S Integer)
+intColumnDelta tn cn = latticeState.lasIntColumnDeltas.singular (ix tn).
+  singular (ix cn)
+
+decColumnDelta :: TableName -> ColumnName -> Lens' AnalyzeState (S Decimal)
+decColumnDelta tn cn = latticeState.lasDecColumnDeltas.singular (ix tn).
+  singular (ix cn)
+
+rowReadCount :: TableName -> S RowKey -> Lens' AnalyzeState (S Integer)
+rowReadCount tn sRk = latticeState.lasRowsRead.singular (ix tn).
+  symArrayAt sRk.sbv2S
+
+rowWriteCount :: TableName -> S RowKey -> Lens' AnalyzeState (S Integer)
+rowWriteCount tn sRk = latticeState.lasRowsWritten.singular (ix tn).
+  symArrayAt sRk.sbv2S
+
+cellEnforced
+  :: TableName
+  -> ColumnName
+  -> S RowKey
+  -> Lens' AnalyzeState (S Bool)
+cellEnforced tn cn sRk = latticeState.lasCellsEnforced.singular (ix tn).
+  singular (ix cn).symArrayAt sRk.sbv2S
+
+cellWritten
+  :: TableName
+  -> ColumnName
+  -> S RowKey
+  -> Lens' AnalyzeState (S Bool)
+cellWritten tn cn sRk = latticeState.lasCellsWritten.singular (ix tn).
+  singular (ix cn).symArrayAt sRk.sbv2S
+
+intCell
+  :: TableName
+  -> ColumnName
+  -> S RowKey
+  -> S Bool
+  -> Lens' AnalyzeState (S Integer)
+intCell tn cn sRk sDirty = latticeState.lasTableCells.singular (ix tn).scIntValues.
+  singular (ix cn).symArrayAt sRk.sbv2SFrom (fromCell tn cn sRk sDirty)
+
+boolCell
+  :: TableName
+  -> ColumnName
+  -> S RowKey
+  -> S Bool
+  -> Lens' AnalyzeState (S Bool)
+boolCell tn cn sRk sDirty = latticeState.lasTableCells.singular (ix tn).scBoolValues.
+  singular (ix cn).symArrayAt sRk.sbv2SFrom (fromCell tn cn sRk sDirty)
+
+stringCell
+  :: TableName
+  -> ColumnName
+  -> S RowKey
+  -> S Bool
+  -> Lens' AnalyzeState (S String)
+stringCell tn cn sRk sDirty = latticeState.lasTableCells.singular (ix tn).scStringValues.
+  singular (ix cn).symArrayAt sRk.sbv2SFrom (fromCell tn cn sRk sDirty)
+
+decimalCell
+  :: TableName
+  -> ColumnName
+  -> S RowKey
+  -> S Bool
+  -> Lens' AnalyzeState (S Decimal)
+decimalCell tn cn sRk sDirty = latticeState.lasTableCells.singular (ix tn).scDecimalValues.
+  singular (ix cn).symArrayAt sRk.sbv2SFrom (fromCell tn cn sRk sDirty)
+
+timeCell
+  :: TableName
+  -> ColumnName
+  -> S RowKey
+  -> S Bool
+  -> Lens' AnalyzeState (S Time)
+timeCell tn cn sRk sDirty = latticeState.lasTableCells.singular (ix tn).scTimeValues.
+  singular (ix cn).symArrayAt sRk.sbv2SFrom (fromCell tn cn sRk sDirty)
+
+ksCell
+  :: TableName
+  -> ColumnName
+  -> S RowKey
+  -> S Bool
+  -> Lens' AnalyzeState (S KeySet)
+ksCell tn cn sRk sDirty = latticeState.lasTableCells.singular (ix tn).scKsValues.
+  singular (ix cn).symArrayAt sRk.sbv2SFrom (fromCell tn cn sRk sDirty)
+
+symArrayAt
+  :: forall array k v
+   . (SymWord k, SymWord v, SymArray array)
+  => S k -> Lens' (array k v) (SBV v)
+symArrayAt (S _ symKey) = lens getter setter
+  where
+    getter :: array k v -> SBV v
+    getter arr = readArray arr symKey
+
+    setter :: array k v -> SBV v -> array k v
+    setter arr = writeArray arr symKey
+
+
+nameAuthorized
+  :: (MonadReader r m, HasAnalyzeEnv r, MonadError AnalyzeFailure m)
+  => S KeySetName
+  -> m (S Bool)
+nameAuthorized sKsn = fmap sansProv $
+  readArray <$> view ksAuths <*> (_sSbv <$> resolveKeySet sKsn)
+
+resolveKeySet
+  :: (MonadReader r m, HasAnalyzeEnv r, MonadError AnalyzeFailure m)
+  => S KeySetName
+  -> m (S KeySet)
+resolveKeySet sKsn = fmap (withProv $ fromNamedKs sKsn) $
+  readArray <$> view keySets <*> pure (_sSbv sKsn)
